wip: refactor kinesis event source to use EventSourceListener class
