+++ tests/integration/awslambda/test_lambda_integration.py
@@ -779,8 +779,12 @@ class TestKinesisSource:
         logs_client,
         lambda_su_role,
     ):
+        import localstack.services.awslambda.event_source_listeners.event_source_listener_kinesis
+
         function_name = f"lambda_func-{short_uid()}"
         stream_name = f"test-foobar-{short_uid()}"
+        num_records_per_batch = 10
+        num_batches = 2
 
         create_lambda_function(
             handler_file=TEST_LAMBDA_PARALLEL_FILE,
@@ -800,61 +804,55 @@ class TestKinesisSource:
             EventSourceArn=stream_arn,
             FunctionName=function_name,
             StartingPosition="TRIM_HORIZON",
-            BatchSize=10,
+            BatchSize=num_records_per_batch,
         )
 
         stream_summary = kinesis_client.describe_stream_summary(StreamName=stream_name)
-        assert 1 == stream_summary["StreamDescriptionSummary"]["OpenShardCount"]
-        num_events_kinesis = 10
-        # assure async call
-        start = time.perf_counter()
-        kinesis_client.put_records(
-            Records=[
-                {"Data": '{"batch": 0}', "PartitionKey": f"test_{i}"}
-                for i in range(0, num_events_kinesis)
-            ],
-            StreamName=stream_name,
-        )
-        assert (time.perf_counter() - start) < 1  # this should not take more than a second
-        kinesis_client.put_records(
-            Records=[
-                {"Data": '{"batch": 1}', "PartitionKey": f"test_{i}"}
-                for i in range(0, num_events_kinesis)
-            ],
-            StreamName=stream_name,
-        )
+        assert stream_summary["StreamDescriptionSummary"]["OpenShardCount"] == 1
 
-        def get_events():
+        for i in range(num_batches):
+            start = time.perf_counter()
+            kinesis_client.put_records(
+                Records=[
+                    {"Data": json.dumps({"record_id": j}), "PartitionKey": f"test_{i}"}
+                    for j in range(0, num_records_per_batch)
+                ],
+                StreamName=stream_name,
+            )
+            assert (time.perf_counter() - start) < 1  # this should not take more than a second
+
+        def get_lambda_invocation_events():
             events = get_lambda_log_events(
                 function_name, regex_filter=r"event.*Records", logs_client=logs_client
             )
             assert len(events) == 2
             return events
 
-        events = retry(get_events, retries=30)
+        invocation_events = retry(get_lambda_invocation_events, retries=30)
 
-        def assertEvent(event, batch_no):
-            assert 10 == len(event["event"]["Records"])
+        for i in range(num_batches):
+            event = invocation_events[i]
+            assert len(event["event"]["Records"]) == num_records_per_batch
+            actual_record_ids = []
+            for record in event["event"]["Records"]:
+                assert "eventID" in record
+                assert "eventSourceARN" in record
+                assert "eventSource" in record
+                assert "eventVersion" in record
+                assert "eventName" in record
+                assert "invokeIdentityArn" in record
+                assert "awsRegion" in record
+                assert "kinesis" in record
+                record_data = record["kinesis"]["Data"]
+                # TODO: I think the record data should actually be b64 encoded
+                actual_record_id = json.loads( record_data )["record_id"]
+                actual_record_ids.append(actual_record_id)
 
-            assert "eventID" in event["event"]["Records"][0]
-            assert "eventSourceARN" in event["event"]["Records"][0]
-            assert "eventSource" in event["event"]["Records"][0]
-            assert "eventVersion" in event["event"]["Records"][0]
-            assert "eventName" in event["event"]["Records"][0]
-            assert "invokeIdentityArn" in event["event"]["Records"][0]
-            assert "awsRegion" in event["event"]["Records"][0]
-            assert "kinesis" in event["event"]["Records"][0]
+            actual_record_ids.sort()
+            assert actual_record_ids == [i for i in range(num_records_per_batch)]
 
-            assert {"batch": batch_no} == json.loads(
-                base64.b64decode(event["event"]["Records"][0]["kinesis"]["data"]).decode(
-                    config.DEFAULT_ENCODING
-                )
-            )
 
-        assertEvent(events[0], 0)
-        assertEvent(events[1], 1)
-
-        assert (events[1]["executionStart"] - events[0]["executionStart"]) > 5
+        assert (invocation_events[1]["executionStart"] - invocation_events[0]["executionStart"]) > 5
 
 
 def _check_mapping_state(lambda_client, uuid):
